Revert "fix(lyrics): bounds-check start_ms_skipping_artefact + split Phase 4 tests"
This reverts commit c1734c0b32723cc399cee5323c7c694ae9421818.
